refactor: Split mini parser and interpreter test cases into separate modules
